Suggested changes

+++ src/test/utils.py
@@ -1,10 +1,14 @@
 import os
-from lib.common.analyser import Analyser
-from lib.common.get_module import get_module
-import shutil
 import json
+import shutil
+from types import SimpleNamespace as Ns
+from pathlib import Path
+from lib.common.storage import LocalStorage
+from lib.common.get import get_module
 
 TEMP_ELEMENT_DIR = "/mtriage/media/test_official"
+TMP_DIR = Path("/tmp")
+STUB_PATHS = Ns(imagejpg="/mtriage/src/test/etype_stubs/image.jpeg",)
 
 
 def scaffold_empty(
@@ -16,12 +20,14 @@ def scaffold_empty(
     Only a single selector should be passed, as derived elements are nested within a selector pass. To create multiple
     selector passes, call this function multiple times.
     """
-    derived_dir = f"{TEMP_ELEMENT_DIR}/{selector}/{Analyser.DERIVED_EXT}"
+    derived_dir = f"{TEMP_ELEMENT_DIR}/{selector}/{LocalStorage.ANALYSED_EXT}"
     if not os.path.exists(derived_dir):
         os.makedirs(derived_dir)
 
     for element in elements:
-        element_dir = f"{TEMP_ELEMENT_DIR}/{selector}/{Analyser.DATA_EXT}/{element}"
+        element_dir = (
+            f"{TEMP_ELEMENT_DIR}/{selector}/{LocalStorage.RETRIEVED_EXT}/{element}"
+        )
         if not os.path.exists(element_dir):
             os.makedirs(element_dir)
         if selector_txt is not None:
@@ -29,14 +35,16 @@ def scaffold_empty(
                 ftxt.write(selector_txt)
         if len(analysers) > 0:
             for analyser in analysers:
-                analyser_dir = f"{TEMP_ELEMENT_DIR}/{selector}/{Analyser.DERIVED_EXT}/{analyser}/{element}"
+                analyser_dir = f"{TEMP_ELEMENT_DIR}/{selector}/{LocalStorage.ANALYSED_EXT}/{analyser}/{element}"
                 if not os.path.exists(analyser_dir):
                     os.makedirs(analyser_dir)
 
 
 def get_element_path(selname, elementId, analyser=None):
     middle_insert = (
-        Analyser.DATA_EXT if analyser is None else f"{Analyser.DERIVED_EXT}/{analyser}"
+        LocalStorage.RETRIEVED_EXT
+        if analyser is None
+        else f"{LocalStorage.ANALYSED_EXT}/{analyser}"
     )
     return f"{TEMP_ELEMENT_DIR}/{selname}/{middle_insert}/{elementId}"
 
@@ -48,7 +56,11 @@ def scaffold_elementmap(elements=[]):
 
 
 def cleanup():
-    shutil.rmtree(TEMP_ELEMENT_DIR)
+    if Path(TEMP_ELEMENT_DIR).exists():
+        shutil.rmtree(TEMP_ELEMENT_DIR)
+    if TMP_DIR.exists():
+        shutil.rmtree(TMP_DIR)
+        TMP_DIR.mkdir()
 
 
 def listOfDictsEqual(l1, l2):
@@ -85,3 +97,8 @@ def list_files(startpath):
         subindent = " " * 4 * (level + 1)
         for f in files:
             print("{}{}".format(subindent, f))
+
+
+def ltemp():
+    """ Primarily for pdb debugging """
+    list_files(TEMP_ELEMENT_DIR)
